// src/app/models/Game.tsx
export interface GameData {
    About_the_Game_Title: string;
    Category: string;
    Comments_Count: string;
    Date: string;
    Developer: string;
    Genre: string;
    Image_URL: string;
    Magnet_URL: string;
    Minimum_Requirements: string;
    NFO: string;
    Post_Details: string;
    Post_URL: string;
    Publisher: string;
    Recommended_Requirements: string;
    Release_Date: string;
    Size: string; // Assuming Size is part of the data structure, if not, remove this line
    Title: string;
    id:any;
    Torrent_URL: string;
    crack_only:string;
  }
  